make informationLogger a class
